% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnehm.R
\name{ada.dnehm}
\alias{ada.dnehm}
\title{A function to estimate DNEHM parameters via adaptive lasso. This is used internally by adalasso.dnehm.}
\usage{
ada.dnehm(data_for_dnehm, node, time, event, cascade, covariates,
  threshold = 0, a = -8)
}
\arguments{
\item{time}{A character string name of the variable that gives the time, in integers}

\item{event}{A character string name of the variable that gives the binary 0/1 indicator of event occurrence.}

\item{covariates}{character vector of covariate names to include in the dnehm, excluding the intercept.}

\item{threshold}{An integer such that an edge variable for node pair i/j will not be constructed if j did not experience more than 'threshold' events after i experienced them.}

\item{a}{A non-negative numeric value that models the exponential decay of sender influence. The effect of a previous event experienced by a diffusion tie sender on the log odds of an event at time t is gamma*exp(-exp(a)*(t-source_time)), where source_time is the time the sender experienced the event.}

\item{eha_data}{A dataframe that includes one observation for each node at risk of experiencing the event during each at-risk time point in each cascade. Note, it is assumed that each node can experience an event in each cascade once, at most.}

\item{A}{character string name of the variable that gives the node id}

\item{A}{character string name of the variable that gives the cascade id}
}
\description{
A function to estimate DNEHM parameters via adaptive lasso. This is used internally by adalasso.dnehm.
}
