% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neha.R
\name{bolasso.neha}
\alias{bolasso.neha}
\title{A function for NEHA estimation with bolasso (Bach 2008) selection.}
\usage{
bolasso.neha(eha_data, node, time, event, cascade, covariates = NULL,
  a = -8, estimate.a = T, n_jobs = 1, data_only = F)
}
\arguments{
\item{eha_data}{A dataframe that includes one observation for each node at risk of experiencing the event during each at-risk time point in each cascade. Note, it is assumed that each node can experience an event in each cascade once, at most.}

\item{node}{A character string name of the variable that gives the node id}

\item{time}{A character string name of the variable that gives the time, in integers}

\item{event}{A character string name of the variable that gives the binary 0/1 indicator of event occurrence.}

\item{cascade}{A character string name of the variable that gives the cascade id}

\item{covariates}{character vector of covariate names to include in the neha, excluding the intercept.}

\item{a}{A non-negative numeric value that models the exponential decay of sender influence. The effect of a previous event experienced by a diffusion tie sender on the log odds of an event at time t is gamma*exp(-exp(a)*(t-source_time)), where source_time is the time the sender experienced the event.}

\item{estimate.a}{A logical value indicating whether or not to estimate the value of 'a'.}

\item{n_jobs}{Integer, number of jobs to run in parallel. -1 means using all processors.}
}
\value{
A list with the following objects.
\itemize{
  \item neha.estimate - glm object giving the final neha estimates.
  \item a.estimate - estimate of a.
  \item formula.neha - formula needed to run final neha specification.
  \item data.for.neha - data frame that includes the edge variables required to run final neha specification.
}
}
\description{
A function for NEHA estimation with bolasso (Bach 2008) selection.
}
\examples{
\dontrun{
# Simulation study of the precision and recall of NEHA
# basic data parameters
cascades <- 100
nodes <- 10
times <- 20
nties <- 25

# generate dataframe
time <- sort(rep(1:times,nodes))
node <- as.character(rep(1:nodes,times))
intercept <- rep(1,length(time))
covariate <- runif(length(time))-2
data_for_sim <- data.frame(time,node,intercept,covariate,stringsAsFactors=F)

# regression parameters
beta <- cbind(c(-2,.25))
rownames(beta) <- c("intercept","covariate")
# generate network effects
possible_ties <- rbind(t(combn(1:nodes,2)),t(combn(1:nodes,2))[,c(2,1)])
possible_ties <- paste(possible_ties[,1],possible_ties[,2],sep="_")
ties <- sample(possible_ties,nties)
gamma <- cbind(exp(rnorm(nties)/2))
rownames(gamma) <- ties
# initiate simulated data object
simulated_data <- NULL
# generate the data one cascade at a time
for(c in 1:cascades){
  simulated_cascade <- simulate_neha_discrete(x=data_for_sim,node="node",time="time",beta=beta,gamma=gamma,a=-8)
  simulated_cascade <- data.frame(simulated_cascade,cascade=c,stringsAsFactors=F)
  simulated_data <- rbind(simulated_data,simulated_cascade)
}

bolasso.results <- bolasso.neha(simulated_data,node="node",time="time",event="event",cascade="cascade",covariates="covariate",a=-8)

bolasso.coefs <- names(coef(bolasso.results[[1]]))[-(1:length(beta))]
recall <- mean(is.element(paste('e',ties,sep=""),bolasso.coefs))
precision <- mean(is.element(bolasso.coefs,paste('e',ties,sep="")))
recall
precision
summary(bolasso.results)
}
}
\references{
Bach, Francis R. "Bolasso: model consistent lasso estimation through the bootstrap." In Proceedings of the 25th international conference on Machine learning, pp. 33-40. ACM, 2008.
}
