% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neha.R
\name{neha_geta}
\alias{neha_geta}
\title{A function to estimate NEHA parameters using greedy edge addition}
\usage{
neha_geta(eha_data, node, time, event, cascade, ncore = 2)
}
\arguments{
\item{eha_data}{A dataframe that includes one observation for each node at risk of experiencing the event during each at-risk time point in each cascade. Note, it is assumed that each node can experience an event in each cascade once, at most.}

\item{node}{A character string name of the variable that gives the node id}

\item{time}{A character string name of the variable that gives the time, in integers}

\item{event}{A character string name of the variable that gives the binary 0/1 indicator of event occurrence.}

\item{cascade}{A character string name of the variable that gives the cascade id}

\item{ncore}{integer indicating the number of cores to use in parallel computation.}
}
\description{
A function to estimate NEHA parameters using greedy edge addition
}
